# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ayden's Therapy Services Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get portal settings (public)
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update portal settings (staff only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /settings/my-hours:
    get:
      operationId: getMyHours
      tags: [settings]
      summary: Get the signed-in therapist's own hours
      responses:
        "200":
          description: Therapist schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TherapistHours"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMyHours
      tags: [settings]
      summary: Update the signed-in therapist's own hours
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TherapistHoursUpdate"
      responses:
        "200":
          description: Updated therapist schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TherapistHours"
        "400":
          description: Invalid schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    DayHours:
      type: object
      required: [open, close, closed]
      properties:
        open:
          type: string
          description: "HH:MM (24h)"
        close:
          type: string
          description: "HH:MM (24h)"
        closed:
          type: boolean
    HolidayHour:
      type: object
      required: [name, date, closed, open, close]
      properties:
        name:
          type: string
        date:
          type: string
          description: "MM-DD format for recurring annual holidays"
        closed:
          type: boolean
        open:
          type: string
        close:
          type: string
    ClosedDate:
      type: object
      required: [date, reason]
      properties:
        date:
          type: string
          description: "YYYY-MM-DD"
        reason:
          type: string
    Settings:
      type: object
      required: [acceptingClients, sessionRequestsOpen, officeHours, holidayHours, closedDates, therapistHours]
      properties:
        acceptingClients:
          type: boolean
        sessionRequestsOpen:
          type: boolean
          description: If false, the booking form is disabled/shows fully booked
        officeHours:
          type: object
          description: Map of day abbreviation to DayHours
          properties:
            mon: { $ref: "#/components/schemas/DayHours" }
            tue: { $ref: "#/components/schemas/DayHours" }
            wed: { $ref: "#/components/schemas/DayHours" }
            thu: { $ref: "#/components/schemas/DayHours" }
            fri: { $ref: "#/components/schemas/DayHours" }
            sat: { $ref: "#/components/schemas/DayHours" }
            sun: { $ref: "#/components/schemas/DayHours" }
        holidayHours:
          type: array
          items:
            $ref: "#/components/schemas/HolidayHour"
        closedDates:
          type: array
          items:
            $ref: "#/components/schemas/ClosedDate"
        siteName:
          type: string
        siteTagline:
          type: string
        logoUrl:
          type: string
          description: Optional HTTP or HTTPS image URL used for the portal logo
        heroTitle:
          type: string
        heroDescription:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        accentColor:
          type: string
        therapistHours:
          type: array
          items:
            $ref: "#/components/schemas/TherapistHours"
    TherapistHours:
      type: object
      required: [name, officeHours]
      properties:
        name:
          type: string
        email:
          type: string
        officeHours:
          type: object
          properties:
            mon: { $ref: "#/components/schemas/DayHours" }
            tue: { $ref: "#/components/schemas/DayHours" }
            wed: { $ref: "#/components/schemas/DayHours" }
            thu: { $ref: "#/components/schemas/DayHours" }
            fri: { $ref: "#/components/schemas/DayHours" }
            sat: { $ref: "#/components/schemas/DayHours" }
            sun: { $ref: "#/components/schemas/DayHours" }
    TherapistHoursUpdate:
      type: object
      required: [officeHours]
      properties:
        officeHours:
          type: object
          properties:
            mon: { $ref: "#/components/schemas/DayHours" }
            tue: { $ref: "#/components/schemas/DayHours" }
            wed: { $ref: "#/components/schemas/DayHours" }
            thu: { $ref: "#/components/schemas/DayHours" }
            fri: { $ref: "#/components/schemas/DayHours" }
            sat: { $ref: "#/components/schemas/DayHours" }
            sun: { $ref: "#/components/schemas/DayHours" }
...
    SettingsUpdate:
      type: object
      properties:
        acceptingClients:
          type: boolean
        sessionRequestsOpen:
          type: boolean
        officeHours:
          type: object
          properties:
            mon: { $ref: "#/components/schemas/DayHours" }
            tue: { $ref: "#/components/schemas/DayHours" }
            wed: { $ref: "#/components/schemas/DayHours" }
            thu: { $ref: "#/components/schemas/DayHours" }
            fri: { $ref: "#/components/schemas/DayHours" }
            sat: { $ref: "#/components/schemas/DayHours" }
            sun: { $ref: "#/components/schemas/DayHours" }
        holidayHours:
          type: array
          items:
            $ref: "#/components/schemas/HolidayHour"
        closedDates:
          type: array
          items:
            $ref: "#/components/schemas/ClosedDate"
        siteName:
          type: string
        siteTagline:
          type: string
        logoUrl:
          type: string
          maxLength: 2048
          description: Optional HTTP or HTTPS image URL used for the portal logo
        heroTitle:
          type: string
        heroDescription:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        accentColor:
          type: string
...