# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Ayden's Therapy Services Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ClosedDate } from './closedDate';
import type { HolidayHour } from './holidayHour';
import type { SettingsOfficeHours } from './settingsOfficeHours';
import type { TherapistHours } from './therapistHours';

export interface Settings {
  acceptingClients: boolean;
  /** If false, the booking form is disabled/shows fully booked */
  sessionRequestsOpen: boolean;
  /** Map of day abbreviation to DayHours */
  officeHours: SettingsOfficeHours;
  holidayHours: HolidayHour[];
  closedDates: ClosedDate[];
  siteName?: string;
  siteTagline?: string;
  /** Optional HTTP or HTTPS image URL used for the portal logo */
  logoUrl?: string;
  heroTitle?: string;
  heroDescription?: string;
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  therapistHours: TherapistHours[];
}