# lib/api-zod/src/generated/types/therapistHours.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Ayden's Therapy Services Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { TherapistHoursOfficeHours } from './therapistHoursOfficeHours';

export interface TherapistHours {
  name: string;
  email?: string;
  officeHours: TherapistHoursOfficeHours;
}