# lib/api-zod/src/generated/types/therapistHoursUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Ayden's Therapy Services Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { TherapistHoursUpdateOfficeHours } from './therapistHoursUpdateOfficeHours';

export interface TherapistHoursUpdate {
  officeHours: TherapistHoursUpdateOfficeHours;
}